Refactor: Organize codebase into subdirectories
- Created new subdirectories: core, events, services, ui, utils within dispyplus.
- Moved existing files into these new subdirectories based on their functionality.
- Renamed some files for clarity (e.g., Dispyplus.py to bot.py, event_*.py to events/*.py, ui.py to ui/views.py, etc.).
- Updated all import statements across the library and examples to reflect the new structure.
- Ensured __init__.py files correctly export necessary components for the public API.
- Synchronized version number in setup.py and dispyplus/__init__.py.
