+++ dispyplus/core/enums.py
