+++ dispyplus/events/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'events' directory a Python package.
