--- dispyplus/other.py
@@ -1,18 +1,0 @@
-# This file is intentionally left blank for now as its contents have been moved.
-# It may be used for other miscellaneous utilities in the future or removed.
-
-# Potential future imports or utilities could go here.
-# For example, common type aliases or small helper functions that don't fit
-# into more specific modules.
-
-# from typing import TypeVar, Generic # Example of keeping common type hints
-# T = TypeVar('T')
-
-# Currently, all specific classes (EnhancedContext, EnhancedView, InteractionType, etc.)
-# have been moved to their respective new files:
-# - context.py for EnhancedContext
-# - ui_components.py for EnhancedView and related UI building blocks
-# - enums.py for InteractionType
-# - ui.py already exists for more complete UI views like ConfirmationView.
-
-pass
